+++ src/aux-records/Utils.ts
@@ -1,5 +1,4 @@
-import { fromByteArray, toByteArray } from 'base64-js';
-import _, { omitBy, padStart, sortBy, StringChain } from 'lodash';
+import _, { omitBy, padStart, sortBy } from 'lodash';
 import { sha256, hmac } from 'hash.js';
 import { PUBLIC_READ_MARKER } from '@casual-simulation/aux-common';
 import axios from 'axios';