+++ src/aux-server/aux-web/aux-auth/site/HasAccountCard/HasAccountCard.vue
@@ -8,9 +8,32 @@
                         <p>You are not logged in. What do you want to do?</p>
                     </div>
                 </div>
-                <p>
-                    <router-link to="privacy-policy" target="_blank">Privacy Policy</router-link>
-                </p>
+                <div class="policies-grid">
+                    <div>
+                        <p>
+                            <a target="_blank" href="/privacy-policy">Privacy Policy</a>
+                        </p>
+                        <p>
+                            <a target="_blank" href="/code-of-conduct">Code of Conduct</a>
+                        </p>
+                        <p>
+                            <a target="_blank" href="/terms">Terms of Service</a>
+                        </p>
+                    </div>
+                    <div class="spacer"></div>
+                    <div class="policy-cert">
+                        <a
+                            style="display: inline-block; width: 70px; height: 70px"
+                            target="_blank"
+                            href="https://cert.privo.com/#/companies/casualsimulation"
+                        >
+                            <img
+                                src="https://privohub.privo.com/files/images/PRIVO_Cert/KPAS_C2V_104_4_72.png"
+                                alt="COPPA Safe Harbor Certification - Kids Privacy Assured by PRIVO"
+                            />
+                        </a>
+                    </div>
+                </div>
             </md-card-content>
             <md-card-actions>
                 <md-button