+++ src/aux-server/aux-web/shared/MonacoHelpers.ts
@@ -76,6 +76,7 @@ import {
 import { union, sortBy } from 'lodash';
 import { propertyInsertText } from './CompletionHelpers';
 import {
+    BotCalculationContext,
     del,
     edit,
     edits,
@@ -83,6 +84,7 @@ import {
     isModule,
     mergeVersions,
     preserve,
+    SHOW_SCRIPT_ISSUES,
     TagEditOp,
 } from '@casual-simulation/aux-common/bots';
 import { Color } from '@casual-simulation/three';
@@ -108,13 +110,11 @@ import {
 let worker: Worker;
 
 export function setup() {
-    worker = new TypescriptWorker();
-
     // Tell monaco how to create the web workers
     (<any>self).MonacoEnvironment = {
         getWorker: function (moduleId: string, label: string) {
             if (label === 'typescript' || label === 'javascript') {
-                return worker;
+                return (worker = new TypescriptWorker());
             } else if (label === 'html') {
                 return new HtmlWorker();
             } else if (label === 'css') {
@@ -267,11 +267,57 @@ export function watchSimulation(
             }
         }
     }
+    function getShowScriptIssues(
+        context: BotCalculationContext,
+        bot: Bot
+    ): boolean {
+        const value = calculateBooleanTagValue(
+            context,
+            bot,
+            SHOW_SCRIPT_ISSUES,
+            false
+        );
+        return value;
+    }
+
+    function updateShowScriptIssues(bot: Bot) {
+        const showScriptIssues = getShowScriptIssues(null, bot);
+        if (showScriptIssues) {
+            monaco.languages.typescript.javascriptDefaults.setDiagnosticsOptions(
+                {
+                    noSemanticValidation: false,
+                    noSyntaxValidation: false,
+                }
+            );
+            monaco.languages.typescript.typescriptDefaults.setDiagnosticsOptions(
+                {
+                    noSemanticValidation: false,
+                    noSyntaxValidation: false,
+                }
+            );
+        } else {
+            monaco.languages.typescript.javascriptDefaults.setDiagnosticsOptions(
+                {
+                    noSemanticValidation: true,
+                    noSyntaxValidation: false,
+                }
+            );
+            monaco.languages.typescript.typescriptDefaults.setDiagnosticsOptions(
+                {
+                    noSemanticValidation: true,
+                    noSyntaxValidation: false,
+                }
+            );
+        }
+    }
 
     sub.add(
         userBotTagsChanged(simulation).subscribe((change) => {
             if (change.tags.has('theme')) {
                 updateTheme(change.bot);
+            }
+            if (change.tags.has(SHOW_SCRIPT_ISSUES)) {
+                updateShowScriptIssues(change.bot);
             }
         })
     );
@@ -417,7 +463,7 @@ export function watchSimulation(
                     addDefinitionsForPortalBot(
                         data.portalId,
                         data.botId,
-                        monaco.languages.typescript.javascriptDefaults
+                        monaco.languages.typescript.typescriptDefaults
                     );
                 })
             )
@@ -429,12 +475,12 @@ export function watchSimulation(
     addDefinitionsForPortalBot(
         'auth',
         'botId',
-        monaco.languages.typescript.javascriptDefaults
+        monaco.languages.typescript.typescriptDefaults
     );
 
     sub.add(
         simulation.watcher.stateUpdated.subscribe(async (update) => {
-            worker.postMessage({
+            worker?.postMessage({
                 __type: 'state',
                 simId: simulation.id,
                 update,
@@ -689,7 +735,7 @@ function registerEditorActionsForSimulation(simulation: Simulation) {
 export function addDefinitionsForPortalBot(
     portalId: string,
     botId: string,
-    defaults: typeof monaco.languages.typescript.javascriptDefaults
+    defaults: typeof monaco.languages.typescript.typescriptDefaults
 ) {
     const extraLibs = defaults.getExtraLibs();
     const libs: { filePath: string; content: string }[] = [];